Update the Task Manager in the Google ADK sample to be reusable (#318)
This commit decouples the `TaskManager` in
`samples/python/agents/google_adk` to be reusable by all Google ADK
samples.

**Note:** This could potentially be refactored into a Google ADK
specific `common` directory.
